adding rendering chart from data
