+++ src/pages/TasksPage/TasksPage.jsx
@@ -29,7 +29,7 @@ import { tasksSelectors } from '../../redux/tasks';
 
 import { currentSprintOperations } from '../../redux/current-sprint';
 
-const {getSprintsItems } = sprintsSelectors;
+const { getSprintsItems } = sprintsSelectors;
 
 const {
   getCurrentSprintTitle,
@@ -47,7 +47,6 @@ export default function TasksPage() {
   const sprintStartDate = useSelector(getCurrentSprintStartDate);
   const sprintDuration = useSelector(getCurrentSprintDuration);
   const visibleTasks = useSelector(tasksSelectors.getVisibleTasks);
-  const allTasks = useSelector(tasksSelectors.getTasks);
   const array = new Array(sprintDuration).fill('');
   const sprintDates = array.map((item, ind) => {
     const date = dayjs(sprintStartDate).add(ind, 'day').format('DD.MM.YYYY');
@@ -57,65 +56,6 @@ export default function TasksPage() {
   const [isCreateSprintModalShown, setCreateSprintModalIsShown] =
     useState(false);
   const [isCreateTaskModalShown, setCreateTaskModalIsShown] = useState(false);
-
-  // ======== CHART LOGIC ========
-  // console.log(allTasks);
-  const totalPlannedHours = allTasks.reduce(
-    (totalPlannedHours, task) => totalPlannedHours + task.plannedHours,
-    0,
-  );
-  const leg = totalPlannedHours / sprintDuration;
-
-  // Array for red chart line
-  const plannedHoursLeftPerDay = new Array(sprintDuration + 1)
-    .fill(0)
-    .reduce((acc, _, ind) => {
-      const value = Math.ceil(totalPlannedHours - leg * ind);
-      acc.push(value);
-      return acc;
-    }, []);
-  console.log(plannedHoursLeftPerDay);
-
-  const totalSpentHoursPerDay = new Array(sprintDuration)
-    .fill(0)
-    .reduce((acc, _, ind) => {
-      const HoursPerDayObjectsByDay = allTasks.map(
-        task => task.hoursPerDay[ind],
-      );
-
-      const totalHoursSpentForAllTasksToday = HoursPerDayObjectsByDay.reduce(
-        (total, obj) => total + obj.hoursSpent,
-        0,
-      );
-      acc.push(totalHoursSpentForAllTasksToday);
-      return acc;
-    }, []);
-
-  // Array for blue chart line
-  const factHoursLeftByPerDay = new Array(sprintDuration + 1)
-    .fill(0)
-    .reduce((acc, _, ind) => {
-      let value = 0;
-      const i = ind - 1;
-
-      if (acc.length > 0) {
-        const number = acc.length - 1;
-        // console.log(number);
-        const previousHoursLeft = acc[number];
-        value = Math.ceil(previousHoursLeft - totalSpentHoursPerDay[i]);
-        ind !== 0 && acc.push(value);
-
-        return acc;
-      }
-      ind === 0 && acc.push(totalPlannedHours);
-
-      return acc;
-    }, []);
-  console.log(factHoursLeftByPerDay);
-
-  // DatesArray for chart
-  console.log(sprintDates);
-  // ======== END OF CHART LOGIC ========
 
   /*Create sprint*/
   const toggleCreateSprintModal = useCallback(() => {
@@ -221,10 +161,10 @@ export default function TasksPage() {
             </Modal>
           )}
         </div>
-        {/* {visibleTasks.length > 3 ? <ChartModalContainer /> : ''} */}
       </div>
       <div className={st.chart_wrapper}>
-        <ChartModalContainer />
+        {visibleTasks.length >= 3 ? <ChartModalContainer /> : ''}
+        {/* <ChartModalContainer /> */}
       </div>
     </>
   );
